refactor: pull OnboardingScript types out into different files
